Test the current viewscreen
The bug: if the embark-anyone script is executed on any viewscreen other than the embark
viewscreen, it aborts with a stack trace.

This bugfix makes it cleanly abort with a reasonably-descriptive error message.

This was found while diagnosing Issue #5509, but is not related.

Minimal changes to the script.

Note on the code change: by moving the function addCivToEmbarkList() inside the function
embarkAnyone(), addCivToEmbarkList() cannot execute unless embarkAnyone() has at least
passed its safety check.